+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.2.0
+
+-   Implement toggle button
+
 # v0.1.0
 
 -   First release